rename Banner.tsx
